requirements.txt
django
djangorestframework
django-extensions
drf-yasg